improved test bench code

+++ controller/pid.go
@@ -5,19 +5,19 @@ import (
 )
 
 type Pid struct {
-    kp, ki, kd float32
-    alpha float32
-    prevError float32
-    integral float32
-    derivative float32
-    integralBounds [2]float32
+    kp, ki, kd float64
+    alpha float64
+    prevError float64
+    integral float64
+    derivative float64
+    integralBounds [2]float64
 }
 
-func NewPid(kp, ki, kd, alpha float32, integralBounds [2]float32) *Pid {
+func NewPid(kp, ki, kd, alpha float64, integralBounds [2]float64) *Pid {
     return &Pid{kp, ki, kd, alpha, 0, 0, 0, integralBounds}
 }
 
-func (p *Pid) Compute(err, dt float32) float32 {
+func (p *Pid) Compute(err, dt float64) float64 {
     p.integral += err * dt
     p.integral = utils.Clip(p.integral, p.integralBounds[0], p.integralBounds[1])
